requirements.txt
appnope==0.1.0
astroid==2.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.0
decorator==4.3.0
diffimg==0.1.7
docopt==0.6.2
idna==2.7
imutils==0.4.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipython-tikzmagic==0.1.1
isort==4.3.4
jedi==0.12.1
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.2
mccabe==0.6.1
more-itertools==4.2.0
networkx==2.1
numpy==1.14.5
pandas==0.23.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.3
Pygments==2.2.0
pylint==2.0.0
pyparsing==2.2.0
Pyphen==0.9.4
pytesseract==0.2.2
pytest==3.6.1
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
repoze.lru==0.7
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
textstat==0.4.1
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
yarg==0.1.9
